agent_template/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-template"
version = "1.0.0"
description = "Template for Customer Experience Management Agents"
requires-python = ">=3.11"
dependencies = [
    # LangChain and LangGraph
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",

    # A2A Protocol (Google A2A)
    "a2a-sdk>=0.1.0",

    # gRPC
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",

    # HTTP clients
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Async support
    "anyio>=4.0.0",

    # Configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",

    # Observability
    "structlog>=24.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-httpx>=0.41b0",

    # Redis for state management
    "redis>=5.0.0",

    # MCP
    "mcp>=1.0.0",

    # Utilities
    "orjson>=3.9.0",
    "tenacity>=8.2.0",  # Retry with backoff
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
